screenshot every two seconds is working.
I altered the timestep code so I can use some functions outside of the class.
I had to make the buffer passed in mutable because we need to be able to flip the rows before saving. Have to pass the width and the height so we know the dims for saving.

auto with std::invocable was breaking in msvc. works with template<>