I think I have all the fields working to update the tile
And refresh the map. Though I don't really have a way to save the changes.
I kind of want to allow editing the source position with out changing the tile. I used a vector for that in my old version. I basically made a complete copy of the tiles and only made changes to the `back();` and I drew the tile in the `front();`