fix spacing with checkboxes put in check boxes getting errors.
I might need to store the front and back maps in variable and not a vector. so I can put the history in the vector but keep current values in a variable that isn't going to move around. OR I can use std::list as that won't move around either.