To eliminate empty tiles it checks each pixel in a tile for a 0 alpha. Then it'll remove the ones with all 0s.
https://github.com/ocornut/imgui/issues/3740
I was trying to eliminate empty tiles. That was kind of a headache. I forgot to divide by 16. So I spent hours like, "why isn't this working?!"  Had to write a little code to show what is in the tiles afterwards. Though they can't be saved as is. Still have to calculate the source x, y, texture_page, and offset the x,y as the center of the drawn image is 0,0 (atleast on most images).