port MapHistory into sfml version and anything it needs.
fixed bug in copy_back where it was returning the wrong reference.
revised remove_duplicate to use a while loop and undo. So if there is more than one duplicate we will be able to remove them all.
tweaked undo to be able to skip redo.
this also gives us redo and undo all and redo all.