added details of tiles you click on.
moved some of the strings to gui labels.
I also wanted to show details of tiles. so i saved what you click on to a vector and show it in the control panel so you can open up a tile and look at it.

I'm not sure if scale_window(); is needed anymore.