Update conanfile.py
set the cmake version to 3.31.6 and I added my program's name to the conanfile.py also set the generator for cmake for windows since visual studio was defaulting to Ninja and vscode was not I couldn't get anything to work when using Ninja. I donno why. So now it's set to visual studio 17 2022 for windows.