fix saving image with save_image_pbo
fix set_uniforms
moved some code from another function into generate_texture so should be able to use the code to draw.
Some of the async code doesn't need to be there.
I'm wondering if the framebuffer needs to survive I'm not sure we might run into issues if the frame buffer goes out of scope. or if i just like write over the old textures while it's running. I'll see what happens. The original code did keep the framebuffer in the future so i might need to do that.