make sure textures get loaded.
fallback_textures function now will check to see if textures are loaded if not it disables loading png textures and reloads.
i had to make some values mutable and some functions const because saving the displayed image as png would run functions to make sure the textures are loaded first.