Framebuffer isn't resizing properly when we change fields.
it is running actually but it's skipping over where it assigns the new dimensions to the framebuffer. will need to go into debug mode as the code is optimized away in runtime heh.
oh i bet i see what the problem is lol && instead of || so it was only resizing when both dims are different.
  if (m_render_framebuffer->width() != spec.width && m_render_framebuffer->height() != spec.height)
gosh i wonder how long that issue was there probably is in the current build.