m_regenerate_items
Added regenerate. Which populates the entries for the current item with the PupuIDs. You can use this if you have no entries as well to give ya a starting point. Added this because I have no undo button. lol. I wanted to be able to atleast restart from where I started.