Allow searching and let the user select items to rename (with checkmarks).
https://github.com/Sebanisu/Field-Map-Editor/issues/185

add table_move struct to hold pointer to parent and child toml tables
add m_select_for_fix_names  to track check marks
removed const from toml::table pointers because I need to be able to modify the tables.
we put the buttons and checkboxes in the same line with the treenodeex.